Fix drawable type

+++ objects/IPlayer.h
@@ -78,7 +78,7 @@ namespace alt
 		virtual void SetWeather(uint32_t weather) = 0;
 		virtual void Kick(StringView reason = "Kicked") = 0;
 		virtual alt::Cloth GetClothes(uint8_t component) const = 0;
-		virtual void SetClothes(uint8_t component, uint8_t drawable, uint8_t texture, uint8_t palette) = 0;
+		virtual void SetClothes(uint8_t component, uint16_t drawable, uint8_t texture, uint8_t palette) = 0;
 #else
 		virtual bool IsTalking() const = 0;
 		virtual float GetMicLevel() const = 0;
